removing duplication tests
