+++ rnote-ui/src/workspacebrowser/filerow/action/duplicate.rs
@@ -1,6 +1,3 @@
-#[cfg(test)]
-mod tests;
-
 use std::ffi::{OsStr, OsString};
 use std::path::PathBuf;
 
@@ -161,4 +158,27 @@ fn remove_dup_word(source_stem: &PathBuf) -> PathBuf {
 
     let removed_dup_suffix = re.replace(&source_stem, "$rest").to_string();
     PathBuf::from(removed_dup_suffix)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_remove_dup_suffix() {
+        // test on filename without ".dup" in name
+        let normal = PathBuf::from("normal_file.txt");
+        let normal_expected = normal.clone();
+        assert_eq!(normal_expected, remove_dup_word(&normal));
+
+        // test with ".dup" name
+        let normal_dup = PathBuf::from("normal_file.dup.txt");
+        let normal_dup_expected = PathBuf::from("normal_file.txt");
+        assert_eq!(normal_dup_expected, remove_dup_word(&normal_dup));
+
+        // test with ".dup1" which means, that a duplicated file has been duplicated
+        let normal_dup1 = PathBuf::from("normal_file.dup1.txt");
+        let normal_dup1_expected = PathBuf::from("normal_file.txt");
+        assert_eq!(normal_dup1_expected, remove_dup_word(&normal_dup1));
+    }
 }
